ageGenderEmotion-master/requirements.txt
streamlit
numpy
opencv-python
keras
matplotlib
pillow
